+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Require NodeJS 22
+
 ## [5.0.0] – 2026-01-24
 
 ### Removed
@@ -72,6 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This CHANGELOG file
 - Allow querying for historical data using `fetchHistoric` and `fetchHistoric90d`
 
+[unreleased]: https://github.com/qqilihq/ecb-euro-exchange-rates/compare/v5.0.0...HEAD
 [5.0.0]: https://github.com/qqilihq/ecb-euro-exchange-rates/compare/v4.0.1...v5.0.0
 [4.0.1]: https://github.com/qqilihq/ecb-euro-exchange-rates/compare/v4.0.0...v4.0.1
 [4.0.0]: https://github.com/qqilihq/ecb-euro-exchange-rates/compare/v3.0.0...v4.0.0